Extract editor to a dedicated component and separate data from componenets
