+++ assets/js/pages/ObjectivePage/index.tsx
@@ -1,13 +1,13 @@
 import React from 'react';
 import { useTranslation } from 'react-i18next';
-import { useQuery, gql } from '@apollo/client';
+import { useQuery, gql, useApolloClient, ApolloClient } from '@apollo/client';
 import { useParams } from 'react-router-dom';
 
 import Avatar from '../../components/Avatar';
 import PageTitle from '../../components/PageTitle';
 import KeyResults from './KeyResults';
 import Projects from './Projects';
-import Editor from './Editor';
+import Editor, { EditorMentionSearchFunc } from '../../components/Editor';
 
 const GET_OBJECTIVE = gql`
   query GetObjective($id: ID!) {
@@ -17,24 +17,35 @@ const GET_OBJECTIVE = gql`
       description
 
       owner {
-        full_name
+        fullName
         title
       }
     }
   }
 `;
 
+const SEARCH_PEOPLE = gql`
+  query SearchPeople($query: String!) {
+    searchPeople(query: $query) {
+      id
+      fullName
+      title
+    }
+  }
+`;
+
 interface Person {
-  full_name: string;
+  fullName: string;
   title: string;
+  id: string;
 }
 
 function Champion({person} : {person: Person}) : JSX.Element {
   return (
     <div className="mt-4 flex gap-2 items-center">
-      <Avatar person_full_name={person.full_name} />
+      <Avatar person_full_name={person.fullName} />
       <div>
-        <div className="font-bold">{person.full_name}</div>
+        <div className="font-bold">{person.fullName}</div>
         <div className="text-sm">{person.title}</div>
       </div>
     </div>
@@ -52,9 +63,30 @@ export async function ObjectivePageLoader(apolloClient : any, {params}) {
   return {};
 }
 
+function PeopleSuggestions(client : ApolloClient<any>) : EditorMentionSearchFunc {
+  return ({query}) => {
+    return new Promise((resolve) => {
+      client
+        .query({ query: SEARCH_PEOPLE, variables: { query } })
+        .then(({ data }) => {
+          resolve(
+            data.searchPeople.map((person : Person) => ({
+              id: person.id,
+              label: person.fullName,
+            })));
+        })
+        .catch((err : any) => {
+          console.log(err);
+        });
+    });
+  }
+}
+
 export function ObjectivePage() {
   const { t } = useTranslation();
   const { id } = useParams();
+  const client = useApolloClient();
+  const peopleSearch = PeopleSuggestions(client);
 
   if (!id) return <p>Unable to find objective</p>;
 
@@ -76,7 +108,11 @@ export function ObjectivePage() {
       <KeyResults objectiveID={id} />
       <Projects objectiveID={id} />
 
-      <Editor />
+      <Editor
+        title={t("objectives.write_an_update.title")}
+        placeholder={t("objectives.write_an_update.placeholder")}
+        peopleSearch={peopleSearch}
+        />
     </div>
   )
 }
